IMS-VT: Change parameter type of data usage
Change Data Usage methods to take parameters of type long
instead of int

CRs-Fixed: 747694

+++ tests/src/com/android/server/telecom/testapps/TestVideoProvider.java
@@ -171,7 +171,7 @@ public class TestVideoProvider extends Connection.VideoProvider {
     @Override
     public void onRequestCallDataUsage() {
         log("Requested call data usage");
-        int dataUsageKb = (10 *1024) + random.nextInt(50 * 1024);
+        long dataUsageKb = (10 *1024) + random.nextInt(50 * 1024);
         changeCallDataUsage(dataUsageKb);
     }
